move GetDeviceByVolumeID to dedicated file
